+++ Changes.md
@@ -23,6 +23,8 @@
    - Every `photo.Descriptor` should define an image for the `photo.Original` key, which should provide a full-size, original image.
    - The `photo.FromImage` option has been added for use with the `ResizeImage` method of `photo.Service`, which selects which `photo.Image` associated with a `photo.Descriptor` should be used as the source image for resizing.
  * New utility function `unsplash.IDFromURL()` added which will give you the photo ID from an Unsplash photo URL.
+ * :hammer: Fix: Multiple total chapter references will be resolved to chapter ranges. For example, if you parse "Ps. 12-13", the resolver will correctly return "Psalms 12-13" instead of "Psalms 12:1-13:6" as it would have before.
+ * :hammer: Fix: Psalms singular handling has been special cased so that references to a single verse can show something like "Psalm 12" or "Psalm 12:1-3" rather than "Psalms 12" or "Psalms 12:1-3" as it would have before.
 
 ## 0.5.1  2024-02-20